Create README.md
Hi Colleagues,
I am a Chartered Engineer (Structural), have an interest in calculation automation and graphics in this field, but as above would like to broaden horizons in general development / programming, robotics, graphics and data science. For the latter currently looking to establish knowledge in data acquisition for analysis, API's web scraping  etc.